requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.3.3
bcrypt==3.1.6
blinker==1.4
cached-property==1.5.1
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
coveralls==1.5.1
cryptography==2.6.1
dbus-python==1.2.12
distlib==0.3.0
distro==1.3.0
Django==2.1.2
django-cors-middleware==1.2.0
django-extensions==1.7.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docker==3.4.1
docker-compose==1.21.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
factory-boy==2.11.1
Faker==0.9.2
fasteners==0.12.0
filelock==3.0.12
future==0.16.0
httplib2==0.11.3
idna==2.7
importlib-metadata==1.5.0
isort==4.3.21
jsonschema==2.6.0
keyring==18.0.1
keyrings.alt==3.1.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
macaroonbakery==1.2.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
oauthlib==2.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==6.1.0
protobuf==3.6.1
psycopg2==2.7.7
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.34.0
PyJWT==1.6.4
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==1.9.0+ubuntu1.3
python-dateutil==2.7.3
python-debian==0.1.36
pytz==2018.5
pyxdg==0.25
PyYAML==5.1.2
reportlab==3.5.23
requests==2.19.1

